clean up map fly animation code
- dont expose some functions on map.ui object
- use event triggers instead
- move animate code to exported function in util module
